Split files into valid extensions
